tests/wpt/tests/html/canvas/offscreen/the-canvas-state/2d.state.saverestore.underflow.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.state.saverestore.underflow</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.state.saverestore.underflow</h1>
<p class="desc">restore() with an empty stack has no effect</p>


<script>
var t = async_test("restore() with an empty stack has no effect");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  for (var i = 0; i < 16; ++i)
      ctx.restore();
  ctx.lineWidth = 0.5;
  ctx.restore();
  _assertSame(ctx.lineWidth, 0.5, "ctx.lineWidth", "0.5");
  t.done();

});
</script>
